+++ packages/vexide-core/src/competition.rs
@@ -7,6 +7,7 @@ use core::{
     cell::UnsafeCell,
     future::{Future, IntoFuture},
     marker::{PhantomData, PhantomPinned},
+    ops::ControlFlow,
     pin::{pin, Pin},
     task::{self, Poll},
 };
@@ -175,9 +176,9 @@ pub const fn updates() -> CompetitionUpdates {
 /// A future which delegates to different futures depending on the current competition mode.
 /// I.e., a tiny async runtime specifically for writing competition programs.
 #[pin_project]
-pub struct Competition<
+pub struct CompetitionRuntime<
     Shared: 'static,
-    Error,
+    Return,
     MkConnected,
     MkDisconnected,
     MkDisabled,
@@ -185,15 +186,15 @@ pub struct Competition<
     MkDriver,
 > where
     MkConnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisconnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisabled:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkAutonomous:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDriver:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
 {
     // Functions to generate tasks for each mode.
     mk_connected: MkConnected,
@@ -221,7 +222,7 @@ pub struct Competition<
     ///   We rely on lifetime parametricity of the `mk_*` functions for this (see the HRTBs above).
     /// - This field MUST come before `shared`, as struct fields are dropped in declaration order.
     #[allow(clippy::type_complexity)]
-    task: Option<Pin<Box<dyn Future<Output = Result<(), Error>> + 'static>>>,
+    task: Option<Pin<Box<dyn Future<Output = ControlFlow<Return>> + 'static>>>,
 
     /// A cell containing the data shared between all tasks.
     ///
@@ -232,7 +233,7 @@ pub struct Competition<
     ///         during task creation.
     shared: UnsafeCell<Shared>,
 
-    /// Keep `self.current` in place while `self.current` references it.
+    /// Keep `self.shared` in place while `self.task` references it.
     _pin: PhantomPinned,
 }
 
@@ -244,21 +245,29 @@ enum CompetitionRuntimePhase {
     Mode(CompetitionMode),
 }
 
-impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver> Future
-    for Competition<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
+impl<Shared, Return, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver> Future
+    for CompetitionRuntime<
+        Shared,
+        Return,
+        MkConnected,
+        MkDisconnected,
+        MkDisabled,
+        MkAutonomous,
+        MkDriver,
+    >
 where
     MkConnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisconnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisabled:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkAutonomous:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDriver:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
 {
-    type Output = Result<(), Error>; // TODO: switch to `!` when stable.
+    type Output = Return;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Self::Output> {
         let mut this = self.as_mut().project();
@@ -291,9 +300,9 @@ where
         }
 
         if let Some(Poll::Ready(res)) = this.task.as_mut().map(|task| task.as_mut().poll(cx)) {
-            // If a task returned an error, then we stop the competition lifecycle and resolve with an error.
-            if let Err(err) = res {
-                return Poll::Ready(Err(err));
+            // If a task says to break out of the competition lifecycle, then we pass the return value up.
+            if let ControlFlow::Break(val) = res {
+                return Poll::Ready(val);
             }
 
             // Reset the current task to nothing, since we're done with the previous task.
@@ -340,24 +349,24 @@ where
     }
 }
 
-impl<Shared, Error>
-    Competition<
+impl<Shared, Return>
+    CompetitionRuntime<
         Shared,
-        Error,
-        DefaultMk<Shared, Error>,
-        DefaultMk<Shared, Error>,
-        DefaultMk<Shared, Error>,
-        DefaultMk<Shared, Error>,
-        DefaultMk<Shared, Error>,
+        Return,
+        DefaultMk<Shared, Return>,
+        DefaultMk<Shared, Return>,
+        DefaultMk<Shared, Return>,
+        DefaultMk<Shared, Return>,
+        DefaultMk<Shared, Return>,
     >
 {
     /// Create a typed builder for a competition runtime with the given `shared` data.
     /// The default tasks simply do nothing, so you do not need to supply them if you don't want to.
-    pub const fn builder(shared: Shared) -> CompetitionBuilder<Shared, Error> {
-        fn default_mk<Shared, Error>(
+    pub const fn builder(shared: Shared) -> CompetitionBuilder<Shared, Return> {
+        fn default_mk<Shared, Return>(
             _: &mut Shared,
-        ) -> Pin<Box<dyn Future<Output = Result<(), Error>>>> {
-            Box::pin(async { Ok(()) })
+        ) -> Pin<Box<dyn Future<Output = ControlFlow<Return>>>> {
+            Box::pin(async { ControlFlow::Continue(()) })
         }
 
         CompetitionBuilder {
@@ -367,23 +376,23 @@ impl<Shared, Error>
             mk_disabled: default_mk,
             mk_autonomous: default_mk,
             mk_driver: default_mk,
-            _error: PhantomData,
+            _return: PhantomData,
         }
     }
 }
 
-type DefaultMk<Shared, Error> =
-    for<'t> fn(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>;
+type DefaultMk<Shared, Return> =
+    for<'t> fn(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>;
 
 /// A typed builder for [`Competition`] instances.
 pub struct CompetitionBuilder<
     Shared,
-    Error,
-    MkConnected = DefaultMk<Shared, Error>,
-    MkDisconnected = DefaultMk<Shared, Error>,
-    MkDisabled = DefaultMk<Shared, Error>,
-    MkAutonomous = DefaultMk<Shared, Error>,
-    MkDriver = DefaultMk<Shared, Error>,
+    Return,
+    MkConnected = DefaultMk<Shared, Return>,
+    MkDisconnected = DefaultMk<Shared, Return>,
+    MkDisabled = DefaultMk<Shared, Return>,
+    MkAutonomous = DefaultMk<Shared, Return>,
+    MkDriver = DefaultMk<Shared, Return>,
 > {
     shared: Shared,
 
@@ -393,14 +402,14 @@ pub struct CompetitionBuilder<
     mk_autonomous: MkAutonomous,
     mk_driver: MkDriver,
 
-    // We're contravariant in the error type.
-    _error: PhantomData<fn(Error)>,
+    // We're invariant in the return type.
+    _return: PhantomData<fn(Return) -> Return>,
 }
 
-impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
+impl<Shared, Return, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
     CompetitionBuilder<
         Shared,
-        Error,
+        Return,
         MkConnected,
         MkDisconnected,
         MkDisabled,
@@ -409,22 +418,29 @@ impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDri
     >
 where
     MkConnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisconnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisabled:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkAutonomous:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDriver:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
 {
     /// Finish the builder, returning a [`Competition`] instance.
     pub fn finish(
         self,
-    ) -> Competition<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
-    {
-        Competition {
+    ) -> CompetitionRuntime<
+        Shared,
+        Return,
+        MkConnected,
+        MkDisconnected,
+        MkDisabled,
+        MkAutonomous,
+        MkDriver,
+    > {
+        CompetitionRuntime {
             mk_connected: self.mk_connected,
             mk_disconnected: self.mk_disconnected,
             mk_disabled: self.mk_disabled,
@@ -440,11 +456,11 @@ where
     }
 }
 
-impl<Shared: 'static, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
+impl<Shared: 'static, Return, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
     IntoFuture
     for CompetitionBuilder<
         Shared,
-        Error,
+        Return,
         MkConnected,
         MkDisconnected,
         MkDisabled,
@@ -453,31 +469,38 @@ impl<Shared: 'static, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomo
     >
 where
     MkConnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisconnected:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDisabled:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkAutonomous:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
     MkDriver:
-        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 't>>,
+        for<'t> FnMut(&'t mut Shared) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 't>>,
 {
-    type Output = Result<(), Error>;
+    type Output = Return;
 
-    type IntoFuture =
-        Competition<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>;
+    type IntoFuture = CompetitionRuntime<
+        Shared,
+        Return,
+        MkConnected,
+        MkDisconnected,
+        MkDisabled,
+        MkAutonomous,
+        MkDriver,
+    >;
 
     fn into_future(self) -> Self::IntoFuture {
         self.finish()
     }
 }
 
-impl<Shared, Error, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
+impl<Shared, Return, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
     CompetitionBuilder<
         Shared,
-        Error,
-        DefaultMk<Shared, Error>,
+        Return,
+        DefaultMk<Shared, Return>,
         MkDisconnected,
         MkDisabled,
         MkAutonomous,
@@ -489,9 +512,11 @@ impl<Shared, Error, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
     pub fn on_connect<Mk>(
         self,
         mk_connected: Mk,
-    ) -> CompetitionBuilder<Shared, Error, Mk, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
+    ) -> CompetitionBuilder<Shared, Return, Mk, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
     where
-        Mk: for<'s> FnMut(&'s mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 's>>,
+        Mk: for<'s> FnMut(
+            &'s mut Shared,
+        ) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 's>>,
     {
         CompetitionBuilder {
             shared: self.shared,
@@ -500,17 +525,17 @@ impl<Shared, Error, MkDisconnected, MkDisabled, MkAutonomous, MkDriver>
             mk_disabled: self.mk_disabled,
             mk_autonomous: self.mk_autonomous,
             mk_driver: self.mk_driver,
-            _error: self._error,
+            _return: self._return,
         }
     }
 }
 
-impl<Shared, Error, MkConnected, MkDisabled, MkAutonomous, MkDriver>
+impl<Shared, Return, MkConnected, MkDisabled, MkAutonomous, MkDriver>
     CompetitionBuilder<
         Shared,
-        Error,
+        Return,
         MkConnected,
-        DefaultMk<Shared, Error>,
+        DefaultMk<Shared, Return>,
         MkDisabled,
         MkAutonomous,
         MkDriver,
@@ -521,9 +546,11 @@ impl<Shared, Error, MkConnected, MkDisabled, MkAutonomous, MkDriver>
     pub fn on_disconnect<Mk>(
         self,
         mk_disconnected: Mk,
-    ) -> CompetitionBuilder<Shared, Error, MkConnected, Mk, MkDisabled, MkAutonomous, MkDriver>
+    ) -> CompetitionBuilder<Shared, Return, MkConnected, Mk, MkDisabled, MkAutonomous, MkDriver>
     where
-        Mk: for<'s> FnMut(&'s mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 's>>,
+        Mk: for<'s> FnMut(
+            &'s mut Shared,
+        ) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 's>>,
     {
         CompetitionBuilder {
             shared: self.shared,
@@ -532,18 +559,18 @@ impl<Shared, Error, MkConnected, MkDisabled, MkAutonomous, MkDriver>
             mk_disabled: self.mk_disabled,
             mk_autonomous: self.mk_autonomous,
             mk_driver: self.mk_driver,
-            _error: self._error,
+            _return: self._return,
         }
     }
 }
 
-impl<Shared, Error, MkConnected, MkDisconnected, MkAutonomous, MkDriver>
+impl<Shared, Return, MkConnected, MkDisconnected, MkAutonomous, MkDriver>
     CompetitionBuilder<
         Shared,
-        Error,
+        Return,
         MkConnected,
         MkDisconnected,
-        DefaultMk<Shared, Error>,
+        DefaultMk<Shared, Return>,
         MkAutonomous,
         MkDriver,
     >
@@ -553,9 +580,11 @@ impl<Shared, Error, MkConnected, MkDisconnected, MkAutonomous, MkDriver>
     pub fn while_disabled<Mk>(
         self,
         mk_disabled: Mk,
-    ) -> CompetitionBuilder<Shared, Error, MkConnected, MkDisconnected, Mk, MkAutonomous, MkDriver>
+    ) -> CompetitionBuilder<Shared, Return, MkConnected, MkDisconnected, Mk, MkAutonomous, MkDriver>
     where
-        Mk: for<'s> FnMut(&'s mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 's>>,
+        Mk: for<'s> FnMut(
+            &'s mut Shared,
+        ) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 's>>,
     {
         CompetitionBuilder {
             shared: self.shared,
@@ -564,19 +593,19 @@ impl<Shared, Error, MkConnected, MkDisconnected, MkAutonomous, MkDriver>
             mk_disabled,
             mk_autonomous: self.mk_autonomous,
             mk_driver: self.mk_driver,
-            _error: self._error,
+            _return: self._return,
         }
     }
 }
 
-impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkDriver>
+impl<Shared, Return, MkConnected, MkDisconnected, MkDisabled, MkDriver>
     CompetitionBuilder<
         Shared,
-        Error,
+        Return,
         MkConnected,
         MkDisconnected,
         MkDisabled,
-        DefaultMk<Shared, Error>,
+        DefaultMk<Shared, Return>,
         MkDriver,
     >
 {
@@ -585,9 +614,11 @@ impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkDriver>
     pub fn while_autonomous<Mk>(
         self,
         mk_autonomous: Mk,
-    ) -> CompetitionBuilder<Shared, Error, MkConnected, MkDisconnected, MkDisabled, Mk, MkDriver>
+    ) -> CompetitionBuilder<Shared, Return, MkConnected, MkDisconnected, MkDisabled, Mk, MkDriver>
     where
-        Mk: for<'s> FnMut(&'s mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 's>>,
+        Mk: for<'s> FnMut(
+            &'s mut Shared,
+        ) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 's>>,
     {
         CompetitionBuilder {
             shared: self.shared,
@@ -596,20 +627,20 @@ impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkDriver>
             mk_disabled: self.mk_disabled,
             mk_autonomous,
             mk_driver: self.mk_driver,
-            _error: self._error,
+            _return: self._return,
         }
     }
 }
 
-impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous>
+impl<Shared, Return, MkConnected, MkDisconnected, MkDisabled, MkAutonomous>
     CompetitionBuilder<
         Shared,
-        Error,
+        Return,
         MkConnected,
         MkDisconnected,
         MkDisabled,
         MkAutonomous,
-        DefaultMk<Shared, Error>,
+        DefaultMk<Shared, Return>,
     >
 {
     /// Use the given function to create a task that runs while the robot is driver controlled.
@@ -617,9 +648,11 @@ impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous>
     pub fn while_driving<Mk>(
         self,
         mk_driver: Mk,
-    ) -> CompetitionBuilder<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, Mk>
+    ) -> CompetitionBuilder<Shared, Return, MkConnected, MkDisconnected, MkDisabled, MkAutonomous, Mk>
     where
-        Mk: for<'s> FnMut(&'s mut Shared) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 's>>,
+        Mk: for<'s> FnMut(
+            &'s mut Shared,
+        ) -> Pin<Box<dyn Future<Output = ControlFlow<Return>> + 's>>,
     {
         CompetitionBuilder {
             shared: self.shared,
@@ -628,76 +661,64 @@ impl<Shared, Error, MkConnected, MkDisconnected, MkDisabled, MkAutonomous>
             mk_disabled: self.mk_disabled,
             mk_autonomous: self.mk_autonomous,
             mk_driver,
-            _error: self._error,
+            _return: self._return,
         }
     }
 }
 
 /// A set of tasks to run when the competition is in a particular mode.
 #[allow(async_fn_in_trait)]
-pub trait CompetitionRobot: Sized {
-    /// The type of errors which can be returned by tasks.
-    type Error;
-
+pub trait Compete: Sized {
     /// Runs when the competition system is connected.
     ///
     /// See [`CompetitionBuilder::on_connect`] for more information.
-    async fn connected(&mut self) -> Result<(), Self::Error> {
-        Ok(())
-    }
+    async fn connected(&mut self) {}
 
     /// Runs when the competition system is disconnected.
     ///
     /// See [`CompetitionBuilder::on_disconnect`] for more information.
-    async fn disconnected(&mut self) -> Result<(), Self::Error> {
-        Ok(())
-    }
+    async fn disconnected(&mut self) {}
 
     /// Runs when the robot is disabled.
     ///
     /// See [`CompetitionBuilder::while_disabled`] for more information.
-    async fn disabled(&mut self) -> Result<(), Self::Error> {
-        Ok(())
-    }
+    async fn disabled(&mut self) {}
 
     /// Runs when the robot is put into autonomous mode.
     ///
     /// See [`CompetitionBuilder::while_autonomous`] for more information.
-    async fn autonomous(&mut self) -> Result<(), Self::Error> {
-        Ok(())
-    }
+    async fn autonomous(&mut self) {}
 
     /// Runs when the robot is put into driver control mode.
     ///
     /// See [`CompetitionBuilder::while_driving`] for more information.
-    async fn driver(&mut self) -> Result<(), Self::Error> {
-        Ok(())
-    }
+    async fn driver(&mut self) {}
 }
 
-/// Extension methods for [`CompetitionRobot`].
-/// Automatically implemented for any type implementing [`CompetitionRobot`].
-pub trait CompetitionRobotExt: CompetitionRobot {
+/// Extension methods for [`Competition`].
+/// Automatically implemented for any type implementing [`Competition`].
+pub trait CompeteExt: Compete {
     /// Build a competition runtime that competes with this robot.
     fn compete(
         self,
-    ) -> Competition<
+    ) -> CompetitionRuntime<
         Self,
-        Self::Error,
-        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 's>>,
-        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 's>>,
-        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 's>>,
-        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 's>>,
-        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 's>>,
+        !,
+        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = ControlFlow<!>> + 's>>,
+        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = ControlFlow<!>> + 's>>,
+        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = ControlFlow<!>> + 's>>,
+        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = ControlFlow<!>> + 's>>,
+        impl for<'s> FnMut(&'s mut Self) -> Pin<Box<dyn Future<Output = ControlFlow<!>> + 's>>,
     > {
-        Competition::builder(self)
-            .on_connect(|s| Box::pin(s.connected()))
-            .on_disconnect(|s| Box::pin(s.disconnected()))
-            .while_disabled(|s| Box::pin(s.disabled()))
-            .while_autonomous(|s| Box::pin(s.autonomous()))
-            .while_driving(|s| Box::pin(s.driver()))
+        #[allow(clippy::unit_arg)]
+        CompetitionRuntime::builder(self)
+            .on_connect(|s| Box::pin(async { ControlFlow::Continue(s.connected().await) }))
+            .on_disconnect(|s| Box::pin(async { ControlFlow::Continue(s.disconnected().await) }))
+            .while_disabled(|s| Box::pin(async { ControlFlow::Continue(s.disabled().await) }))
+            .while_autonomous(|s| Box::pin(async { ControlFlow::Continue(s.autonomous().await) }))
+            .while_driving(|s| Box::pin(async { ControlFlow::Continue(s.driver().await) }))
             .finish()
     }
 }
 
-impl<R: CompetitionRobot> CompetitionRobotExt for R {}
+impl<R: Compete> CompeteExt for R {}