+++ packages/vexide-math/src/lib.rs
@@ -1,6 +1,0 @@
-//! Common mathematical formulas and models implemented for [`vexide`](https://crates.io/crates/vexide).
-
-#![no_std]
-
-pub mod feedforward;
-pub mod pid;