+++ packages/vexide-startup/examples/booting.rs
@@ -6,7 +6,9 @@
 extern crate alloc;
 use alloc::boxed::Box;
 
+use vex_sdk::vexTasksRun;
 use vexide_core::println;
+use vexide_startup::{CodeSignature, ProgramFlags, ProgramOwner, ProgramType};
 
 #[no_mangle]
 extern "Rust" fn main() {
@@ -15,20 +17,25 @@ extern "Rust" fn main() {
         let test_box = Box::new(100);
         vex_sdk::vexDisplayRectFill(0, 0, *test_box, 200);
         println!("Hello, world!");
+        vexTasksRun(); // Flush serial
     }
 }
 
 #[no_mangle]
 #[link_section = ".boot"]
-unsafe extern "C" fn _entry() {
-    unsafe { vexide_startup::program_entry() }
+unsafe extern "C" fn _start() {
+    unsafe { vexide_startup::program_entry::<true>() }
 }
 
-#[link_section = ".cold_magic"]
+#[link_section = ".code_signature"]
 #[used] // This is needed to prevent the linker from removing this object in release builds
-static COLD_HEADER: vexide_startup::ColdHeader = vexide_startup::ColdHeader::new(2, 0, 0);
+static CODE_SIG: CodeSignature = CodeSignature::new(
+    ProgramType::User,
+    ProgramOwner::Partner,
+    ProgramFlags::empty(),
+);
 
 #[panic_handler]
-fn panic(_info: &core::panic::PanicInfo<'_>) -> ! {
+const fn panic(_info: &core::panic::PanicInfo<'_>) -> ! {
     loop {}
 }